Put SignaturePledge into index file
