+++ src/components/Layout/Header/Menu/LoginMenuItem.jsx
@@ -17,7 +17,7 @@ const LoginTitle = ({ userData, userId }) => {
   return (
     <div className={s.loginParentTitle}>
       <AvatarImage className={s.loginParentAvatar} user={userData} />
-      <span class={s.loginParentTitleText}>
+      <span className={s.loginParentTitleText}>
         {userData.username || userData.email}
       </span>
     </div>
